requirements.txt
Django==1.10.5
Pillow==4.0.0
factory_boy
